+++ src/shared/components/IconsList/data.ts
@@ -1,0 +1,17 @@
+import styles from './styles.module.css';
+
+export const iconSources = {
+    student: [
+      { src: "src/assets/images/StudentIcons/loupe.png", className: styles.loupe },
+      { src: "src/assets/images/StudentIcons/toggle.png", className: styles.toggle },
+      { src: "src/assets/images/StudentIcons/cone.png", className: styles.cone },
+      { src: "src/assets/images/StudentIcons/bubble.png", className: styles.bubble },
+      { src: "src/assets/images/StudentIcons/plus.png", className: styles.plus },
+    ],
+    teacher: [
+      { src: "src/assets/images/TeacherIcons/triangle.png", className: styles.triangle },
+      { src: "src/assets/images/TeacherIcons/cube.png", className: styles.cube },
+      { src: "src/assets/images/TeacherIcons/ring.png", className: styles.ring },
+      { src: "src/assets/images/TeacherIcons/spiral.png", className: styles.spiral },
+    ],
+  };
