--- src/widgets/Main/index.tsx
@@ -1,11 +1,0 @@
-import MainBlock from '../../shared/components/MainBlock';
-
-const Main = () => {
-  return (
-    <div>
-      <MainBlock />
-    </div>
-  );
-};
-
-export default Main;
