--- src/widgets/QuestionList/data.ts
@@ -1,40 +1,0 @@
-import { v4 as uuidv4 } from 'uuid';
-
-interface IQuestion {
-  id: string;
-  title: string;
-  content: string;
-}
-
-export const questions: IQuestion[] = [
-  {
-    id: uuidv4(),
-    title: 'Что такое RepetitMe?',
-    content:
-      'RepetitMe - это сервис подбора репетиторов для помощи ученикам в учебе'
-  },
-  {
-    id: uuidv4(),
-    title: 'Как искать репетитора?',
-    content:
-      'Вам необходимо пройти регистрацию и заполнить информацию о себе, после чего выбрать в предложенной системе поиска нужного репетитора по вашему учебному предмету'
-  },
-  {
-    id: uuidv4(),
-    title: 'Есть ли бонусная программа?',
-    content:
-      'В нашем приложении есть бонусная программа. Для ее использования необходимо...'
-  },
-  {
-    id: uuidv4(),
-    title: 'Можно ли вернуть деньги за оплаченное занятие?',
-    content:
-      'Если по какой-то причине занятие не состоялось и вы хотите вернуть деньги, то вам необходимо обратиться к нашей службе поддержки'
-  },
-  {
-    id: uuidv4(),
-    title: 'Есть ли система абонементов?',
-    content:
-      'Разумеется! В нашем приложении есть система абонементов. Для ее оформления необходимо...'
-  }
-];
